Made the plugin compatible with IDEA 223 (2022.3).
The plugin is not compatible wih IDEA 222 or earlier versions anymore. The functionality of the plugin is not going to be extended, so I've made the change to publish the plugin only for IDEA 223.
After that all plugin features will be merged into the main IDEA Kotlin plugin, and will be available starting from IDEA 231.